Break apart budget/misc.js
