+++ .eslintrc.js
@@ -230,13 +230,19 @@ module.exports = {
         './packages/desktop-client/src/components/UpdateNotification.*',
         './packages/desktop-client/src/components/accounts/Header.*',
         './packages/desktop-client/src/components/alerts.*',
+        './packages/desktop-client/src/components/budget/BudgetCategories.*',
+        './packages/desktop-client/src/components/budget/BudgetTotals.*',
+        './packages/desktop-client/src/components/budget/ExpenseGroup.*',
+        './packages/desktop-client/src/components/budget/IncomeGroup.*',
         './packages/desktop-client/src/components/budget/MobileBudget.*',
         './packages/desktop-client/src/components/budget/MobileBudgetTable.*',
         './packages/desktop-client/src/components/budget/MobileTable.*',
         './packages/desktop-client/src/components/budget/MonthCountSelector.*',
         './packages/desktop-client/src/components/budget/MonthPicker.*',
+        './packages/desktop-client/src/components/budget/RenderMonths.*',
+        './packages/desktop-client/src/components/budget/SidebarCategory.*',
+        './packages/desktop-client/src/components/budget/SidebarGroup.*',
         './packages/desktop-client/src/components/budget/constants.*',
-        './packages/desktop-client/src/components/budget/misc.*',
         './packages/desktop-client/src/components/budget/report/BudgetSummary.*',
         './packages/desktop-client/src/components/budget/report/components.*',
         './packages/desktop-client/src/components/budget/rollover/BudgetSummary.*',
