+++ packages/desktop-client/src/components/budget/BudgetTotals.js
@@ -1,0 +1,106 @@
+import React, { memo, useState } from 'react';
+
+import DotsHorizontalTriple from '../../icons/v1/DotsHorizontalTriple';
+import { colors } from '../../style';
+import Button from '../common/Button';
+import Menu from '../common/Menu';
+import View from '../common/View';
+import { Tooltip } from '../tooltips';
+
+import { MONTH_BOX_SHADOW } from './constants';
+import RenderMonths from './RenderMonths';
+import { getScrollbarWidth } from './util';
+
+const BudgetTotals = memo(function BudgetTotals({
+  MonthComponent,
+  toggleHiddenCategories,
+  expandAllCategories,
+  collapseAllCategories,
+}) {
+  const [menuOpen, setMenuOpen] = useState(false);
+  return (
+    <View
+      data-testid="budget-totals"
+      style={{
+        backgroundColor: 'white',
+        flexDirection: 'row',
+        flexShrink: 0,
+        boxShadow: MONTH_BOX_SHADOW,
+        marginLeft: 5,
+        marginRight: 5 + getScrollbarWidth(),
+        borderRadius: '4px 4px 0 0',
+        borderBottom: '1px solid ' + colors.border,
+      }}
+    >
+      <View
+        style={{
+          width: 200,
+          color: colors.n4,
+          justifyContent: 'center',
+          paddingLeft: 15,
+          paddingRight: 5,
+          display: 'flex',
+          flexDirection: 'row',
+          alignItems: 'center',
+          userSelect: 'none',
+          WebkitUserSelect: 'none',
+        }}
+      >
+        <View style={{ flexGrow: '1' }}>Category</View>
+        <Button
+          type="bare"
+          onClick={() => {
+            setMenuOpen(true);
+          }}
+          style={{ color: 'currentColor', padding: 3 }}
+        >
+          <DotsHorizontalTriple
+            width={15}
+            height={15}
+            style={{ color: colors.n5 }}
+          />
+          {menuOpen && (
+            <Tooltip
+              position="bottom-right"
+              width={200}
+              style={{ padding: 0 }}
+              onClose={() => {
+                setMenuOpen(false);
+              }}
+            >
+              <Menu
+                onMenuSelect={type => {
+                  if (type === 'toggleVisibility') {
+                    toggleHiddenCategories();
+                  } else if (type === 'expandAllCategories') {
+                    expandAllCategories();
+                  } else if (type === 'collapseAllCategories') {
+                    collapseAllCategories();
+                  }
+                  setMenuOpen(false);
+                }}
+                items={[
+                  {
+                    name: 'toggleVisibility',
+                    text: 'Toggle hidden',
+                  },
+                  {
+                    name: 'expandAllCategories',
+                    text: 'Expand all',
+                  },
+                  {
+                    name: 'collapseAllCategories',
+                    text: 'Collapse all',
+                  },
+                ]}
+              />
+            </Tooltip>
+          )}
+        </Button>
+      </View>
+      <RenderMonths component={MonthComponent} />
+    </View>
+  );
+});
+
+export default BudgetTotals;
