+++ packages/desktop-client/src/components/budget/CategoryGroupsContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export let CategoryGroupsContext = createContext([]);
